Split LambdaRuntimeClient in two files for easier reading
